Refactor test directory structure
Use a flat structure and use a file per library module.

Use the same singular names as library modules.
